feat(yafti): Make user installation the default
Most apps on the list beyond the first two sections make more sense as
user-preferred apps, less so as core necessities. i.e. I might want to
use Chrome for web development, or chat with friends on Discord, but I
wouldn't want to clutter the apps list of someone else on the same
system that has no need for those.