Add initial dconf settings
Keybindings:
- The default workspace navigation keybindings in GNOME feel a bit out
of place as of GNOME 40; PgUp and PgDown are still used there, but "up"
and "down" don't make as much sense with the workspaces now being
horizontal. Keybinds are changed to be more akin to what elementary OS
uses, and to simplify workspace navigation.
- Window tiling keybinds were adjusted to accomodate the above changes.

Interface tweaks:
- Clock format set to 12h
- Clock shows weekday
- Font antialiasing set to rgba
- Default GTK theme set to adw-gtk3 for better visual consistency with
libadwaita apps

Misc:
- Tap-to-click set to true for touchpads
- New windows are centered instead of the vanilla "smart" placement