refactor: Convert flatpak-manager into a module (#45)
* refactor: Move flatpak-manager and user-setup to modules dir

* fix: Remove just setup from user-setup

* fix: Disable systemd module

We're only enabling the Flatpak-related services, and those are being
converted into a separate module.

* refactor: Use generic names for flatpak setup scripts and services

* refactor: Standardize config file checks in flatpak-setup scripts

* feat: Initial add flatpaks module

* feat: Enable flatpaks module

* fix: Copy system and user scripts separately

* fix: Point to correct directory

* refactor: Move flatpak-setup files back into main usr dirs

Unless this is made a bling module instead, this module is gonna have to
work like yafti in that the necessary files need to be included in
startingpoint directly.

* feat: Create flatpak install and remove lists

* fix: Point system-flatpak-setup to new install and remove files

* feat: Enable flatpak installation and removal for user

* refactor: Move flatpak installation to flatpaks module

* fix: Build install and removal lists correctly

* fix(main): Remove user flatpak installs

Had them there for testing, don't need them to stay.

* chore: Add output for added flatpaks

* style: Make flathub remote name generic

System-wide was flathub-system, per-user was flathub-user. I do like
having it explicit like that myself, but for more general-purpose usage,
it's probably best to keep it generic as just flathub for both. Flatpak
does allow specifying --system or --user already, anyway.

* refactor: Use flatpaks module for kinoite

* fix(dconf): packaging-format-preference: flathub-user -> flathub

Since I'm switching back to using "flathub" as the remote name instead
of specifying -user vs -system