feat: Remove NautilusPreviewer from default apps
It isn't discoverable at all, unless you're coming from macOS and happen
to try hitting spacebar on a file in your home directory.

It uses static permissions instead of the FileChooser portal, and by
default only has access to the home directory, so it doesn't work on
extra internal or external drives. This -could- be configured in a
user-setup script, but going to extra work for something that's also
made redundant by other default apps doesn't feel worth the effort.

I could maybe be convinced to add it back in the future, but for now,
I'm removing it.