build: Rebase onto Fedora 39 (#121)
Switches both Zeliblue and Zeliblue Plasma from the `latest` tag to the
`39` tag. This change already took place for ZeliDeck in
#106.

Also updates the rpm-ostree remove list in Plasma:

gwenview, kcalc, and okular have been
[removed](https://pagure.io/workstation-ostree-config/c/9995a5f61bbaff0744822056737adc894025a6de?branch=f39)
from the upstream Kinoite image and so no longer need to be removed
here.

Kwrite was added back into `kinoite-main`, so it has been added to the
remove list in favor of the Flatpak.