feat: Brewfile updates (#149)
- Removes `webp`: I had originally added it in because I used it as part
of a script to convert screenshots from .png to .webp, but 1) that's my
own use case, not a general use case as is the goal of inclusions in the
Brewfile, and 2) I don't need it anymore anyway as
[Converter/Switcheroo](https://flathub.org/apps/io.gitlab.adhami3310.Converter)
now better handles conversion of multiple images
- Replaces `neofetch` with `fastfetch`: Generates system info much
faster and shows more than `neofetch`
- Removes `htop` in favor of just having `btop`: Makes sense to only
include one `top` alternative. Sticking with btop since it's easier to
use and read, and it looks nicer.