build: Add `stable` and `testing` tags (#259)
`stable` will replace `latest` as the intended tag for most users to be
on. `testing` will be used for experimental changes and images, such as
Zeliblue COSMIC.

The kernel replacement with the fsync kernel that is
currently done in all builds will also be moved into `testing` branches
in another PR, as Bluefin is reverting their own switch to the fsync
kernel and I want to do the same for `stable` images here.