feat(Brewfile): Remove atuin
I've been trialing it for a number of months, but I have never gotten
used to atuin's Up-Arrow behavior compared to the behavior of bash and
fish. As for Ctrl+R search, while atuin is nice for that, we default to
the fish shell already, and fish's is nicer to use than bash.

We'll keep our fish integration for atuin in case users install it
manually, but no longer include atuin in the Brewfile.