refactor(testing): Port Zeliblue and -deck from BlueBuild (#313)
Ports Zeliblue and Zeliblue Deck from the BlueBuild build process over to one derived from Universal Blue's image-template setup. This change is currently restricted to the `testing` tag, and all of the BuildBuild-related files are still present, so reverting this change is as simple as re-adjusting the build-testing GitHub action to point back to the BlueBuild action.