ci: Disable zeliblue-deck builds
I've switched back to SteamOS on my Steam Deck, and as of the SteamOS
`main` branch they're shipping new enough kernel & mesa that it runs on
my gaming PC, too, so I don't currently have any devices running
Zeliblue Deck. I do still like the idea of "SteamOS but with GNOME" and
may return to Zeliblue Deck in the future, but for now I don't really
have much time to work on it anyway.